requirements.txt
pandas
rapidfuzz
fugashi
unidic-lite
pyyaml